Added dependencies, added database creation scripts, added text file to help setup database.
    app.js
Added code to attempt logging in to website.

    package.json files
Added node-supervisor dependency to help with keeping the server up and restarted when encountering an error.
It also helps to restart the server automatically when changing server files when trying to fix or add functions.
Also when run will ignore views and changes to public folder.
To run this properly run supervisor -i public,views app.js

    Creation scripts
Added default for tables creation and user population.

    help setup.txt
Create a text file to help with setting up the database and its tables.