Created batch scripts to help ease tasks, finalized resetting password and logging in.
    triggers.sql
Added triggers to store users prev_passwords so passwords cannot be reused.

    setupDatabase.sql
Added a call to triggers.sql.

    help setup.txt
Added additional instructions to help setup the database.

    app.js
Added app.get to handle requests and render views for resetPassword and main.
Added app.post to handle resetPassword form submission.
Added 404 error message to be used for app.use

    Login To Database.bat
Added batch file to login to database without having to type the command for ease of use.

    Reset Database.bat
Added batch file to reset database to default settings without having to navigate to folders and type the command for ease of use.

    Start Web Server.bat
Added batch file to start the web server without having to navigate to folders for ease of use.

    main.css, main.ejs, main.js
Added code for rendering of main.
This code is reused from a previous project and will be a template to show something for now to test loging in but needs overhaul for this project.