Added logic to handle updating database word list for certain languages.  Added triggers to create tables for words when a language is added.  Added scripts to help organize default database creation.  Added/changed comments for clarification on what the scripts are doing.
    setupDatabase.sql
Changed file call for population to match new organization.

    tables.sql
Added additional echo statements for output cleanliness.
Changed words table to languages and changed its schema.
Words was moved to triggers to be created.
Languages table only lists languages and when they were last updated.

    triggers.sql
Added trigger and procedure to create a table for the words of a language.
This is replacing the words table that was previously in tables.sql.

    dropWordsTables.sql
Added logic to remove all words_* tables that are created when inserting languages.
This is used to reset database to default.

    lang_population.sql
Added script to populate default languages.

    population.sql
Added script to handle all default populations such as lang_population.sql and user_population.sql.

    app.js
Fixed bug with 404 error not reporting username properly of who requested the page.
Changed the app.get for flashCardGame to allow for a parameter for what language the user wants to play the flashCardGame for.
Added logic to check if the words_[language] table had not been updated in 3 months and update the table with the newest word list.

    DatabaseHandler.js
Removed functions from previous project that are not being used.
Added logic to query all available languages.
Added logic to query language word list age.
Added logic to insert words into the database.

    wordsHttpRequest.js
Allowed the function to be exported, so app.js can call the function.

    flashCardGame.ejs
There is currently a bug where this view requests /main/graphics/favicon.png.
Added a link to the correct folder but the bug is still not fixed.

    main.ejs
Changed code from previous project to show links as place holders.
Added the default languages to be shown under Flash Card Games.