Finalized logic for updating and inserting words into the database.  Changed default on language table, and primary key on words tables.sql.
     ***ISSUE***
There is a current problem with the website having duplicate rank numbers but different words.  This is causing some errors on primary key constraints to go off.  At the moment there is no workaround as functionality is still in place.

    tables.sql
Changed languages table to default to NULL instead of NOW().
This allows for initial word population instead of defaulting database to say words are up to date.

    triggers.sql
Changed primary key for words_* tables to just be rank_num not rank_num and word.

    app.js
Finalized logic to handle if words_list for language needs updated or initialized.

    DatabaseHandler.js
Changed insertWord to alterWordList to handle both initializing and updating the words_* tables based on argument supplied.