Refactored update words code for app.js cleanliness.  Changed when update/initialize words code was called to lessen downtime for the user.  Added logic to help with finalizing flash card game.  Added nav menu.  Finalized logic for flashCardGame.
    app.js
Removed call to wordsHttpRequest due to code refactoring.
Added call to updateWordsList due to code refactoring.
Changed logic in app.get for flashCardGame so that only the database is queried prior to rendering the view.  Previously it would check to update database and potentially update the database prior to rendering the view.  This change allows for the user load times to be more efficient.  This logic now calls the new function updateWordsList.
Added check to updateWordsList in app.listen so the server checks if words_* tables need to be updated or initialized on server startup.

    updateWordsList.js
Took previous code in app.get for flashCardGame and refactored it into its own function to be called.
This function checks if needed and then updates or initialized the words_* tables.

    DatabaseHandler.js
Changed languageQuery function to return a promise to easily make it synchronous.
Added function getWordList to handle querying the database for specific languages word table data.

    flashCardGame.ejs
Added navigation menu to allow a user to easily go back to the main menu.
Added text to inform the user to click the flash card to "flip it".
Added id to the logout link to ensure it doesn't conflict with the nave menu.

    flashCardGame.css, nav.css
Changed and added stylizing for above changes in flashCardGame.ejs.

    flashCardGame.js
Re-worked and finalized logic that was pre-emptively created prior to getting real data to test functionality of the flashCardGame logic. The logic was incorrect for real data, but is functioning correctly now.