Fixed an issue with the user table triggers. Added handling of lang_pref for users in server files.
    triggers.sql
There was an issue with updating a certain column that wasn't pass where it would fail updating due to the prev_pass_trig trigger.
I create a separate trigger or updating and inserting.
In the updating trigger made sure it only executed prev_pass_procedure() if the new pass was different than the old pass.

    app.js, DatabaseHandler.js, User.js
Added logic to allow for retreiving, storing, and sending the users lang pref to the view to render only languages the user wants to learn.
View logic still needs to be done.