Finalized logic for showing facts.  Added javascript and css to control facts view logic and style.
    app.js
Removed debug printing lines.

    DatabaseHandler.js
Fixed issue with wrong column name for cultural_page.

    wikipediaAPIRequest.js
Changed ignoreList to a dictonary for faster lookup and added a section to it.
Changed prop style from wiki api query to text instead of wikitext so that it could be directly inserted into the view.

    languageFacts.ejs
Removed debug line.
Added lines to link the new javascript and css files for controlling view logic and styling.
Removed styling that is no longer needed since more specific styling is now in languageFacts.css.

   languageFacts.css
Added styling for the view and the wiki content.

   languageFacts.js
Added logic to add wiki content to the view.
Added logic to control wiki urls to link to appropriate pages.