Added view and logic for searching users.  Added languages spoken to default user population.  Added English to list of languages in lang table.
    lang_population.sql
Added English to default language population

    user_population.sql
Added languages spoken for default users.

    app.js
Added logic for rendering the views searchUsers.ejs.
Added if statement to not update English word list (*This may change in the future if this project goes beyond scholastic purpose*)

    DatabaseHandler.js
Added comments for function purposes that were missing.
Added a function to control searching users for searchUsers.ejs view and for updating it if a user wants to narrow the search.

    main.ejs
Fixed a naming issue with country facts specifying only Italy.
Removed a placeholder and replaced it with the link to searchUsers.ejs view.

    searchUsers.ejs
Added view to handle searching users.

    basicResultsTable.css, search.css, searchUsers.css
Added stylesheets to handle styling for searchUsers.ejs view.

    searchUsers.js
Added logic to handle updating search for searchUsers.ejs view.
*Not finished*