Fixed issue with default user population. Finalized logic for searching users.
    user_population.sql
Fixed an issue where the user_num didn't match users causing the default population to not insert.

    app.js
Added logging for error catch of previous commit view logic.
Added app.get to return results to view for a search of users.

    DatabaseHandler.js
Added check to see if opts exists prior to checking if opts['narrow'] exists.
Fixed indentation of code.

    search.css
Removed code that was part of boiler plate that didn't belong here.

    searchUsers.js
Fixed and finalized logic of ajax call and updateTable so that it would correctly search and update results to the view.

    searchUsers.ejs
Removed creating javascript variable that was not needed to send language to app.gets.
Removed hidden line for search field.