Added app.post to register user to the database and control routing.  Bug fixes to DatabaseHandler.js addNewUser. Changes to register.ejs view.
    app.js
Added app.post to handle logic of registering a user and if there was a duplicate username or email when attempting to register.
Added logic to alphabetize languages when rendering registration view.

    DatabaseHandler.js
Fixed issue when registering it would default to inserting data into the database that would cause a user to need to reset password upon first login.

    loginFailed.ejs
Changed logo class name to represent current project code.  Was leftover from boilerplate code.

    register.ejs
Added logic to help register.js show the user if registration failed due to duplicate username or password.

    register.js
Added file to control logic of showing the user if registration failed due to duplicate username or password.