Made changes to structure of messages table.  Added default population for messages. Added view framework for settings and webmail.  Added app.gets for webmail and settings.  Added logic to get list of messages from database.
    LLUO DB ERD.png, LLUO900 SDS.docx, tables.sql
Changed messages table to be a timestamp for sent_date instead of a date.
Updated this change in the erd and the sds.

    message_population.sql
Added default message population.

    population.sql
Added call to message_population.sql

    app.js
Added app.gets for webmail and settings.
Changed req.params.langPref to req.session.user.langPref since it is in the user class and then less data needs to between each view.

    DatabaseHandler.js
Added function getMessages() to get list of all sent/received messages from the database and return to calling function.

    main.ejs
Added links to the settings view and the webmail view.

    settings.ejs, webmail.ejs
Added basic view framework for settings and webmail.