Fixed bug with database populations. Fixed bug with searching for users not working properly. Added logic and functionality for basic message functionality.
    message_population.sql
File was incorrectly deleting information from the wrong tables.
Fixed it to delete only from messages before repopulating.

    user_population.sql
Re-organized file for better execution output readability.

    webmail.ejs
Changed elements for the message interface from tables to divs.
Added more elements to have better framework, styling and logic handling.

    webmail.js
Added logic to control clicking on inbox and sent and updating the message list.
Added logic to control showing messages when clicking on the message overview.

    webmail.css
Added styling for updated elements of the webmail view.

    app.js
Fixed a bug which led to incorrect searches and routing when searching for users.

    searchUsers.js, searchUsers.ejs
Fixed logic of ajax query going to wrong routing, and added passing variable to ajax query.
This fixed a bug that led to 404 errors when searching for users.