Finalized logic and view for changing settings.
    app.js
Finalized logic in app.get to render settings view.
Added app.post to handle the logic of changing a users settings and displaying to the user whether it was successful or that they couldn't use a previous password.

    DatabaseHandler.js
Created function updateUser to handle logic of updating the users information in the database.
Added instantiation of fname and lname variables to user class in the attemptLogin function.

    User.js
Added lname and fname variables to the class.

    settings.ejs
Added logic and framework to properly render the settings view to the User.js.
Added framework to display to the user when their password meets guidelines.

    settings.css
Added styling for the settings view.

    settings.js
Added logic to animate to the user when they attempted to use a previous password.

    resetPassword.js
Added event handler for submission of the settingsForm on settings.ejs view.