Sent user_num to search view to make sure someone doesn't see themselves in results.
    app.js
Added passing variable user_num to search view.

    DatabaseHandler.js
Added line to queryStrings in getUserList to avoid returning results that include the user.

    searchUsers.js
Fixed issues with incorrect links when displaying users.
Fixed incorrect pathname for ajax query.

    webmail.js
Changed conditional to check if a user had no sent messages to if a user no contacts to allow for them to get to compose if they are a new account from the search users view.